+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.5.1
+
+* Reverts the addition of `onUrlChanged`, which was unintentionally a breaking
+  change.
+
+## 1.5.0
+
+* Added `onUrlChanged` callback to platform callback handler.
+
 ## 1.4.0
 
 * Added `loadFile` and `loadHtml` interface methods.